TunnelsManager: Workaround for macOS Catalina: Show confirmation alert
When reinstating from zero tunnels, macOS will show the VPN prompt
(“WireGuard” Would Like to Add VPN Configurations). Before that prompt,
we show a confirmation alert to make it clear to the user what the VPN
prompt is for.

The text of this prompt is not localized because this is expected to be
a temporary workaround.